+++ react/typings/global.d.ts
@@ -4,8 +4,9 @@ interface Window extends Window {
     clickCode: string
     advertiserCode: string
     key: string
-    iframeDomain: string,
-    _ps_tg: TrackingObject
+    iframeDomain: string
+    cookieScriptDomain: string
+    _ps_tgt: TrackingObject
   }
 }