+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+### [0.3.0]
+- fixed encryption
+- fixed tracking code and conversion code
+- added new order tracking event 
+
 ## [0.2.0]
 - added new variables for general domain use: iframeDomain and cookieScriptDomain
 For Bulgaria, the links are: profitshare.bg